my_project-master/nodeman-training_201905/nodeman-training_201905/requirements.txt
# -*- coding: utf-8 -*-
Django==1.8.11
pymysql==0.6.7
MarkupSafe==0.23
Mako==1.0.4
httplib2==0.9.1
suds==0.4
celery==3.1.18
django-celery==3.2.1
uWSGI==2.0.13.1
requests==2.9.1
djangorestframework==3.6.3
django-jsonfield==1.0.1
django-extensions==1.7.9
rsa==3.4.2
impacket==0.9.15
paramiko==2.0.2
drf-extensions==0.3.1
django-cors-headers==2.1.0
sqlparse==0.2.4
pycrypto==2.6.1